requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1
firebase-admin==6.6.0
google-generativeai==0.8.3
langchain==0.3.13
langchain-community==0.3.13
langchain-google-genai==2.0.7
langchain-huggingface==0.1.2
chromadb==0.5.23
sentence-transformers==3.3.1
pypdf==5.1.0
python-multipart==0.0.20
gradio==5.9.1
requests==2.32.3
